requirements.txt
# Core dependencies
pyyaml>=6.0
click>=8.0
colorama>=0.4.6
jsonschema>=4.17.0
requests>=2.28.0

# Optional for specific skills
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
jinja2>=3.1.0

# Development
pytest>=7.0.0
black>=23.0.0
mypy>=1.0.0